Compute confidence interval/s for the Mann-Whitney parameter

Compute confidence interval/s for the Mann-Whitney parameter

estimate_ci_mannwhitney(mannwhitney_est, cdf_est, pmf_est, treat_prob_est,
  treat_form, out_form, treat, ci, out, alpha, nboot, out_levels, covar,
  out_model)

Arguments

mannwhitney_est
The point estimates for log-odds.

cdf_est
The estimated conditional CDF.

pmf_est
The estimated conditional PMF.

treat_prob_est
Estimated probability of treatments, output from call to estimate_treat_prob.

treat_form
The right-hand side of a regression formula for the working model of treatment probability as a function of covariates

out_form
The right-hand side of a regression formula for the working proportional odds model. NOTE: THIS FORMULA MUST NOT SUPPRESS THE INTERCEPT.

treat
A numeric vector containing treatment status. Should only assume a value 0 or 1.

ci
A vector of characters indicating which confidence intervals should be computed ("bca" and/or "wald")

out
A numeric vector containing the outcomes.

alpha
Confidence intervals have nominal level 1-alpha.

nboot
Number of bootstrap replicates used to compute bootstrap confidence intervals.

out_levels
A numeric vector containing all ordered levels of the outcome.

covar
A data.frame containing the covariates to include in the working proportional odds model.

out_model
Which R function should be used to fit the proportional odds model. Options are "polr" (from the MASS package), "vglm" (from the VGAM package), or "clm" (from the ordinal package).

out_weights
A vector of numeric weights with length equal to the length of out_levels.

Value

List with wald and bca-estimated confidence intervals for the Mann-Whitney parameter.